requirements.txt
psycopg2-binary
python-dotenv
llama-index
llama-index-vector-stores-postgres
llama-index-embeddings-openai
llama-index-llms-openai
sqlalchemy
fastapi
uvicorn
pydantic
